making cycles happy
