+++ server/api/api/api.go
@@ -10,6 +10,19 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
+// m exposing the data abstraction layer
+
+type Manager struct {
+	Datastore      models.Datastore
+	userManagement user.Management
+	Acl            acl.Acl
+	DatabaseName   string
+	HugoHelper     hugo.HugoSiteBuilder
+	OauthHandlers  oauth.Handlers
+	logger         *logrus.Logger
+	insights       event.Insights
+}
+
 func NewManager(
 	datastore models.Datastore,
 	userManagement user.Management,
