+++ server/api/api/label.go
@@ -9,19 +9,20 @@ import (
 	"github.com/freshteapot/learnalist-api/server/api/i18n"
 	"github.com/freshteapot/learnalist-api/server/api/models"
 	"github.com/freshteapot/learnalist-api/server/api/uuid"
+	"github.com/freshteapot/learnalist-api/server/pkg/api"
 	"github.com/labstack/echo/v4"
 )
 
 func (m *Manager) V1PostUserLabel(c echo.Context) error {
 	user := c.Get("loggedInUser").(uuid.User)
-	var input = &HttpLabelInput{}
+	var input = &api.HttpLabelInput{}
 
 	defer c.Request().Body.Close()
 	jsonBytes, _ := ioutil.ReadAll(c.Request().Body)
 
 	err := json.Unmarshal(jsonBytes, input)
 	if err != nil {
-		response := HttpResponseMessage{
+		response := api.HttpResponseMessage{
 			Message: i18n.PostUserLabelJSONFailure,
 		}
 		return c.JSON(http.StatusBadRequest, response)
@@ -35,14 +36,14 @@ func (m *Manager) V1PostUserLabel(c echo.Context) error {
 	case http.StatusCreated:
 		break
 	case http.StatusBadRequest:
-		response := HttpResponseMessage{
+		response := api.HttpResponseMessage{
 			Message: i18n.ValidationLabel,
 		}
 		return c.JSON(http.StatusBadRequest, response)
 	case http.StatusInternalServerError:
 		fallthrough
 	default:
-		response := HttpResponseMessage{
+		response := api.HttpResponseMessage{
 			Message: i18n.InternalServerErrorFunny,
 		}
 		return c.JSON(http.StatusInternalServerError, response)
@@ -57,7 +58,7 @@ func (m *Manager) V1GetUserLabels(c echo.Context) error {
 	labels, err := m.Datastore.GetUserLabels(user.Uuid)
 	if err != nil {
 		// TODO log this
-		response := HttpResponseMessage{
+		response := api.HttpResponseMessage{
 			Message: i18n.InternalServerErrorFunny,
 		}
 		return c.JSON(http.StatusInternalServerError, response)
@@ -69,10 +70,10 @@ func (m *Manager) V1RemoveUserLabel(c echo.Context) error {
 	user := c.Get("loggedInUser").(uuid.User)
 	label := c.Param("label")
 	err := m.Datastore.RemoveUserLabel(label, user.Uuid)
-	response := HttpResponseMessage{}
+	response := api.HttpResponseMessage{}
 	if err != nil {
 		// TODO log this
-		response := HttpResponseMessage{
+		response := api.HttpResponseMessage{
 			Message: i18n.InternalServerErrorFunny,
 		}
 		return c.JSON(http.StatusInternalServerError, response)
