+++ server/api/api/root.go
@@ -3,12 +3,13 @@ package api
 import (
 	"net/http"
 
+	"github.com/freshteapot/learnalist-api/server/pkg/api"
 	"github.com/labstack/echo/v4"
 )
 
 func (m *Manager) V1GetRoot(c echo.Context) error {
 	message := "1, 2, 3. Lets go!"
-	response := HttpResponseMessage{
+	response := api.HttpResponseMessage{
 		Message: message,
 	}
