+++ server/e2e/client.go
@@ -12,8 +12,8 @@ import (
 	"time"
 
 	"github.com/freshteapot/learnalist-api/server/api/alist"
-	"github.com/freshteapot/learnalist-api/server/api/api"
 	"github.com/freshteapot/learnalist-api/server/api/utils"
+	"github.com/freshteapot/learnalist-api/server/pkg/api"
 )
 
 // Generated by curl-to-Go: https://mholt.github.io/curl-to-go
@@ -111,7 +111,7 @@ func (c Client) Register(username string, password string) RegisterResponse {
 	return response
 }
 
-func (c Client) DeleteUser(credentials api.HTTPLoginResponse) (statusCode int, response api.HttpResponseMessage) {
+func (c Client) DeleteUser(credentials api.HttpLoginResponse) (statusCode int, response api.HttpResponseMessage) {
 	url := fmt.Sprintf("%s/api/v1/user/%s", c.getServerURL(), credentials.UserUUID)
 	req, err := http.NewRequest("DELETE", url, nil)
 	req = req.WithContext(context.Background())
