+++ server/e2e/smoke_2_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	"github.com/freshteapot/learnalist-api/server/api/alist"
-	"github.com/freshteapot/learnalist-api/server/api/api"
 	"github.com/freshteapot/learnalist-api/server/e2e"
+	"github.com/freshteapot/learnalist-api/server/pkg/api"
 	. "github.com/onsi/ginkgo"
 	"github.com/stretchr/testify/assert"
 )
